Update and rename server/app.js to api/src/app.js
